Start reorganize files
- put all .cpp files into srcs/ and it's subdirectory
- put all .hpp files into incs/ directory
- modify README, Makefile and main.cpp
