+++ incs/utils.hpp
@@ -1,0 +1,4 @@
+#include "webserv.hpp"
+
+void	*ft_memset(void *s, int c, size_t n);
+void	errMsgErrno(std::string msg);
