+++ srcs/config/Server.cpp
@@ -1,4 +1,4 @@
-#include "Server.hpp"
+#include "../../incs/Server.hpp"
 
 void debug_print(std::map<std::string, Location> x) {
 	for (std::map<std::string, Location>::iterator it = x.begin(); it != x.end(); it++) {
