+++ srcs/config/doParsing.cpp
@@ -1,4 +1,4 @@
-#include "webserv.hpp"
+#include "../../incs/webserv.hpp"
 
 void doServerParsing(std::queue<std::vector<std::string> > &qu, int &line) {
 	if (qu.front().front() != "server")
