+++ srcs/utils/utils.cpp
@@ -1,0 +1,18 @@
+#include "../../incs/utils.hpp"
+
+void	*ft_memset(void *s, int c, size_t n)
+{
+	while (n)
+	{
+		*((unsigned char *)s + n - 1) = (unsigned char)c;
+		n--;
+	}
+	return (s);
+}
+
+void	errMsgErrno(std::string msg) {
+	std::cerr << RED ;
+	std::cerr << msg << "\n";
+	std::cerr << strerror(errno) << "\n";
+	std::cerr << RESET ;
+}
